Declare utf-8 encoding in HTTP headers
While the XML header has already defined the encoding,
this is not necessary. The reason is that before sending
the XML body of the request to the parser, HTTP-server
on eBay side decodes it as a stream of ASCII (by default).

This wrong decoding causes Java XML parser to receive
the text with zero-length symbols if the original
utf-8 contains multi-byte graphemes.

To fix the problem I added the charset to the HTTP
'Content-Type' header, as well as provided the 'Accept'
and 'Accept-Charset' symbols for eBay server to know
how to encode the response.